Move UserPreferences class to Config namespace
